merges bug_fix->trunk
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/aims/aims-free/trunk@69994 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ aimsdata/src/aimsdicom/io/dicomheader.cc
@@ -401,7 +401,7 @@ int DicomHeader::read()
       } 
  
     dimZ = _slices.size() / dimT;
-    if( dimZ * dimT != _slices.size() )
+    if( static_cast<size_t>(dimZ * dimT) != _slices.size() )
     {
       cerr << "DICOM reader warning: slices number does not make a complete "
         << "volume" << std::endl;
